+++ configs/alerter_template.yaml
@@ -1,0 +1,5 @@
+# Rename to alerter.yaml after editing
+
+sender_email: your_gmail_email_address
+receiver_email: receiver_email_address
+password: your_gmail_app_password
