+++ configs/experiment_params.yaml
@@ -1,0 +1,16 @@
+experiment_dataset: mnist
+population_size: 20
+generations: 20
+offspring_proportion: 0.5
+normal_cell_repeats: 1
+substructure_repeats: 1
+parameter_limit: 100000
+complexity_penalty: 0.2
+number_of_runs: 1
+seed: 42
+batch_size: 32
+epochs: 5
+verbose: 1
+optimizer: adam
+loss: categorical_crossentropy
+metrics: [accuracy, mse]
