+++ configs/mutation_phases.yaml
@@ -1,0 +1,14 @@
+rapid_expansion:
+  generation_percentage: 0.5
+  individual_mutation_rate: 1.0
+  mutation_type_rate: [0.8, 0.2, 0.0]
+
+steady_growth:
+  generation_percentage: 0.3
+  individual_mutation_rate: 0.5
+  mutation_type_rate: [0.4, 0.4, 0.2]
+
+local_exploration:
+  generation_percentage: 0.2
+  individual_mutation_rate: 0.6
+  mutation_type_rate: [0.0, 0.5, 0.5]
