+++ configs/network_graphs.yaml
@@ -1,0 +1,34 @@
+input_graph:
+  nodes:
+    Conv2D:
+      filters: 16
+      kernel_size: [3,3]
+      activation: relu
+      padding: same
+    BatchNormalization: {}
+  edges:
+    edge_1: [Conv2D, BatchNormalization]
+
+normal_cell_graph: null
+
+reduction_cell_graph:
+  nodes:
+    AveragePooling2D:
+      pool_size: [2,2]
+      strides: [2,2]
+      padding: valid
+    Conv2D:
+      filters: 16
+      kernel_size: [1,1]
+      activation: relu
+      padding: valid
+  edges:
+    edge_1: [AveragePooling2D, Conv2D]
+
+output_graph:
+  nodes:
+    GlobalAveragePooling2D: {}
+    Dense:
+      activation: softmax
+  edges:
+    edge_1: [GlobalAveragePooling2D, Dense]
