+++ configs/search_space.yaml
@@ -1,0 +1,12 @@
+# Layer/Node options: Conv2D; BatchNormalization; AveragePooling2D; MaxPooling2D; GlobalAveragePooling2D; Dense; SpatialDropout2D; Identity
+
+Conv2D:
+  filters: [8,16,32]
+  kernel_size: [[1,1],[3,3]]
+  activation: [relu]
+  padding: [same]
+
+AveragePooling2D:
+  pool_size: [[3,3]]
+  strides: [[1,1]]
+  padding: [same]
